pain

+++ code/modules/projectiles/unsorted/magic.dm
@@ -12,8 +12,10 @@
 	name = "bolt of death"
 	icon_state = "pulse1_bl"
 
-/obj/projectile/magic/death/on_hit(target, var/mob/living/L)
+/obj/projectile/magic/death/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(ismob(target))
 		var/mob/M = target
 		if(L.anti_magic_check())
@@ -28,8 +30,10 @@
 	damage_type = OXY
 	nodamage = 1
 
-/obj/projectile/magic/resurrection/on_hit(mob/living/carbon/target)
+/obj/projectile/magic/resurrection/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(isliving(target))
 		if(target.anti_magic_check())
 			target.visible_message("<span class='warning'>[src] vanishes on contact with [target]!</span>")
@@ -49,8 +53,10 @@
 	var/inner_tele_radius = 0
 	var/outer_tele_radius = 6
 
-/obj/projectile/magic/teleport/on_hit(mob/target, var/mob/living/L)
+/obj/projectile/magic/teleport/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(ismob(target))
 		var/mob/M = target
 		if(L.anti_magic_check())
@@ -76,8 +82,10 @@
 	nodamage = 1
 	var/list/door_types = list(/obj/structure/simple_door/wood, /obj/structure/simple_door/iron, /obj/structure/simple_door/silver, /obj/structure/simple_door/gold, /obj/structure/simple_door/uranium, /obj/structure/simple_door/sandstone, /obj/structure/simple_door/phoron, /obj/structure/simple_door/diamond)
 
-/obj/projectile/magic/door/on_hit(atom/target)
+/obj/projectile/magic/door/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(istype(target, /obj/machinery/door))
 		OpenDoor(target)
 	else
@@ -291,6 +299,7 @@
 		if(owner)
 			C.ChangeOwner(owner)
 */
+
 /obj/projectile/magic/spellblade
 	name = "blade energy"
 	icon_state = "lavastaff"
@@ -299,14 +308,17 @@
 	sharp = TRUE
 	magic = TRUE
 
-/obj/projectile/magic/spellblade/on_hit(target, var/mob/living/L)
+/obj/projectile/magic/spellblade/on_impact_new(atom/target, impact_flags, def_zone)
+	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
+
 	if(ismob(target))
 		var/mob/M = target
 		if(L.anti_magic_check())
 			M.visible_message("<span class='warning'>[src] vanishes on contact with [target]!</span>")
 			qdel(src)
 			return
-	. = ..()
 
 /obj/projectile/magic/arcane_barrage
 	name = "arcane bolt"
@@ -318,16 +330,19 @@
 	magic = TRUE
 	impact_sounds = 'sound/weapons/barragespellhit.ogg'
 
-/obj/projectile/magic/arcane_barrage/on_hit(target, var/mob/living/L)
+/obj/projectile/magic/arcane_barrage/on_impact_new(atom/target, impact_flags, def_zone)
+	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
+
 	if(ismob(target))
 		var/mob/M = target
 		if(L.anti_magic_check())
 			M.visible_message("<span class='warning'>[src] vanishes on contact with [target]!</span>")
 			qdel(src)
 			return
-	. = ..()
 
-
+/* needs more work
 /obj/projectile/magic/locker
 	name = "locker bolt"
 	icon_state = "locker"
@@ -352,21 +367,24 @@
 	return ..()
 */
 
-/obj/projectile/magic/locker/on_hit(target)
-	if(created)
-		return ..()
-	var/obj/structure/closet/decay/C = new(get_turf(src))
-	if(LAZYLEN(contents))
-		for(var/atom/movable/AM in contents)
+/obj/projectile/magic/locker/on_impact_new(atom/target, impact_flags, def_zone)
+	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
+	if(!created)
+		var/obj/structure/closet/decay/C = new(src)
 		C.update_icon()
-	created = TRUE
-	return ..()
+	if(locker_suck && !(target.atom_flags & (ATOM_ABSTRACT | ATOM_NONWORLD)))
+		if(ismovable(target))
+			var/atom/movable/AM = target
+			AM.forceMove(src)
 
 /obj/projectile/magic/locker/Destroy()
 	locker_suck = FALSE
 	for(var/atom/movable/AM in contents)
 		AM.forceMove(get_turf(src))
 	. = ..()
+*/
 
 /obj/structure/closet/decay
 	breakout_time = 600
@@ -428,16 +446,17 @@
 		chain = caster.Beam(src, icon_state = "lightning[rand(1, 12)]", time = INFINITY, maxdistance = INFINITY)
 	..()
 
-/obj/projectile/magic/aoe/lightning/on_hit(target, var/mob/living/L)
+/obj/projectile/magic/aoe/lightning/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(ismob(target))
 		var/mob/M = target
 		if(L.anti_magic_check())
 			M.visible_message("<span class='warning'>[src] fizzles on contact with [target]!</span>")
-			qdel(src)
-		return blocked
+			return . | PROJECTILE_IMPACT_DELETE
 	tesla_zap(src, zap_range, zap_power)
-	qdel(src)
+	return . | PROJECTILE_IMPACT_DELETE
 
 /obj/projectile/magic/aoe/lightning/Destroy()
 	qdel(chain)
@@ -456,17 +475,21 @@
 	var/exp_flash = 3
 	var/exp_fire = 2
 
-/obj/projectile/magic/aoe/fireball/on_hit(target)
+/obj/projectile/magic/aoe/fireball/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(ismob(target))
 		var/mob/living/M = target
 		if(M.anti_magic_check())
 			visible_message("<span class='warning'>[src] vanishes into smoke on contact with [target]!</span>")
-			return
+			return . | PROJECTILE_IMPACT_DELETE
 		M.take_overall_damage(0,10) //between this 10 burn, the 10 brute, the explosion brute, and the onfire burn, your at about 65 damage if you stop drop and roll immediately
 	var/turf/T = get_turf(target)
 	explosion(T, -1, exp_heavy, exp_light, exp_flash, 0)//, flame_range = exp_fire)
+	return . | PROJECTILE_IMPACT_DELETE
 
+/* requires IMPACT_DELETE_AFTER
 /obj/projectile/magic/aoe/fireball/infernal
 	name = "infernal fireball"
 	exp_heavy = -1
@@ -474,15 +497,14 @@
 	exp_flash = 4
 	exp_fire= 5
 
-/obj/projectile/magic/aoe/fireball/infernal/on_hit(target)
+/obj/projectile/magic/aoe/fireball/infernal/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
-	if(ismob(target))
-		var/mob/living/M = target
-		if(M.anti_magic_check())
-			return
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	var/turf/T = get_turf(target)
 	for(var/i=0, i<50, i+=10)
 		addtimer(CALLBACK(GLOBAL_PROC, GLOBAL_PROC_REF(explosion), T, -1, exp_heavy, exp_light, exp_flash, FALSE, FALSE, exp_fire), i)
+*/
 
 /obj/projectile/magic/nuclear
 	name = "\proper blazing manliness"
@@ -491,7 +513,10 @@
 	var/mob/living/victim = null
 	var/used = 0
 
-/obj/projectile/magic/nuclear/on_hit(target)
+/obj/projectile/magic/nuclear/on_impact_new(atom/target, impact_flags, def_zone)
+	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	if(used)
 		return
 	if(ismob(target))
@@ -500,16 +525,12 @@
 		used = 1
 		visible_message("<span class='danger'>[victim] slams into [target] with explosive force!</span>")
 		explosion(src, 2, 3, 4, -1, TRUE, FALSE, 5)
+		return PROJECTILE_IMPACT_DELETE
 	else
 		used = 1
 		victim.take_overall_damage(30,30)
 		explosion(src, -1, -1, -1, -1, FALSE, FALSE, 5)
-	return
-
-/obj/projectile/magic/nuclear/Destroy()
-	for(var/atom/movable/AM in contents)
-		AM.forceMove(get_turf(src))
-	. = ..()
+		return PROJECTILE_IMPACT_DELETE
 
 //Spellcards
 
@@ -533,8 +554,10 @@
 	damage = 4
 	var/fire_stacks = 4
 
-/obj/projectile/magic/spellcard/book/spark/on_hit(atom/target, blocked = FALSE)
+/obj/projectile/magic/spellcard/book/spark/on_impact_new(atom/target, impact_flags, def_zone)
 	. = ..()
+	if(. & PROJECTILE_IMPACT_FLAGS_SHOULD_ABORT)
+		return
 	var/mob/living/carbon/M = target
 	if(ismob(target))
 		if(M.anti_magic_check())
